Move Name class to shared package
