+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -24,7 +24,7 @@ import seedu.address.model.Model;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
