+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.member.NameContainsKeywordsPredicate;
+import seedu.address.model.shared.NameContainsKeywordsPredicate;
 
 /**
  * Finds and lists all members in address book whose name contains any of the argument keywords.
