+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -15,7 +15,7 @@ import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
