+++ src/main/java/seedu/address/model/member/Member.java
@@ -8,6 +8,7 @@ import java.util.Objects;
 import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.shared.Name;
 import seedu.address.model.tag.Tag;
 
 /**
