+++ src/main/java/seedu/address/model/shared/Name.java
@@ -1,4 +1,4 @@
-package seedu.address.model.member;
+package seedu.address.model.shared;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
