+++ src/main/java/seedu/address/model/shared/NameContainsKeywordsPredicate.java
@@ -1,10 +1,11 @@
-package seedu.address.model.member;
+package seedu.address.model.shared;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.member.Member;
 
 /**
  * Tests that a {@code Member}'s {@code Name} matches any of the keywords given.
