+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -9,7 +9,7 @@ import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
