+++ src/main/java/seedu/address/storage/JsonAdaptedMember.java
@@ -13,7 +13,7 @@ import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
