+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -40,7 +40,7 @@ import seedu.address.logic.commands.AddCommand;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.MemberBuilder;
