+++ src/test/java/seedu/address/testutil/MemberBuilder.java
@@ -6,7 +6,7 @@ import java.util.Set;
 import seedu.address.model.member.Address;
 import seedu.address.model.member.Email;
 import seedu.address.model.member.Member;
-import seedu.address.model.member.Name;
+import seedu.address.model.shared.Name;
 import seedu.address.model.member.Phone;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.util.SampleDataUtil;
